Desktop/PROJECTS/AI/deepseek-document-processor/requirements.txt
streamlit==1.32.0
requests==2.31.0
pdfplumber==0.11.0
python-docx==1.1.0
chardet==5.2.0